BENCH-187-TEST-CONTAINERS Refactored Base Test
