+++ src/test/java/com/answerdigital/answerking/utility/AbstractContainerBaseTest.java
@@ -1,0 +1,26 @@
+package com.answerdigital.answerking.utility;
+
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.MySQLContainer;
+
+@ActiveProfiles("test")
+public abstract class AbstractContainerBaseTest {
+    static final MySQLContainer MY_SQL_CONTAINER;
+
+    static {
+        MY_SQL_CONTAINER =
+                new MySQLContainer<>("mysql:8.0.31")
+                        .withReuse(true);
+        MY_SQL_CONTAINER.start();
+    }
+
+    @DynamicPropertySource
+    private static void overrideProps(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", MY_SQL_CONTAINER::getJdbcUrl);
+        registry.add("spring.datasource.username", MY_SQL_CONTAINER::getUsername);
+        registry.add("spring.datasource.password", MY_SQL_CONTAINER::getPassword);
+        registry.add("spring.datasource.name", MY_SQL_CONTAINER::getDatabaseName);
+    }
+}
